+++ objects/src/batches/mod.rs
@@ -1,0 +1,2 @@
+mod note_tree;
+pub use note_tree::BatchNoteTree;
