+++ objects/src/block/mod.rs
@@ -2,3 +2,5 @@ use super::{Digest, Felt, Hasher, ZERO};
 
 mod header;
 pub use header::BlockHeader;
+mod note_tree;
+pub use note_tree::BlockNoteTree;
