+++ objects/src/notes/mod.rs
@@ -38,9 +38,6 @@ pub use origin::{NoteInclusionProof, NoteOrigin};
 mod script;
 pub use script::NoteScript;
 
-mod tree;
-pub use tree::{BatchOutputNotesTree, BlockOutputNotesTree};
-
 // CONSTANTS
 // ================================================================================================
